Move estimator_checks to testing module as for maintainance
