+++ aeon/utils/tests/test_check_estimator.py
@@ -6,7 +6,7 @@ import pytest
 
 from aeon.testing.mock_estimators import MockClassifier, MockSegmenter
 from aeon.transformations.exponent import ExponentTransformer
-from aeon.utils.estimator_checks import check_estimator
+from aeon.testing.estimator_checks import check_estimator
 
 EXAMPLE_CLASSES = [MockClassifier, MockSegmenter]
